:black_nib: Issue templates: Add template for TeleIRC (closes #119)
This commit adds a new issue template to request TeleIRC chat bridges.
It is a speedier template to request chat bridges than the other more
general-purpose templates. Of note, this template will automatically
assign me and @Tjzabel to issues that request bridges, since we have
commit access to the RITlug TeleIRC infrastructure.

Closes #119.

Signed-off-by: Justin W. Flory <git@jwf.io>